requirements.txt
appium-python-client
pytest
pytest-fixture-config
python-dotenv
feedparser